+++ src/UserSearchView.js
@@ -24,6 +24,7 @@ import {
   SearchAndSortSearchButton as FilterPaneToggle,
 } from '@folio/stripes/smart-components';
 
+import { CREATE_USER_URL } from './constants';
 import filterConfig, { filterConfigWithUserAssignedStatus } from './filterConfig';
 import Filters from './Filters';
 
@@ -58,6 +59,7 @@ class UserSearchView extends React.Component {
     idPrefix: PropTypes.string,
     isMultiSelect: PropTypes.bool,
     onSelectRow: PropTypes.func,
+    onClose: PropTypes.func,
     onSaveMultiple: PropTypes.func,
     onComponentWillUnmount: PropTypes.func,
     queryGetter: PropTypes.func,
@@ -67,6 +69,7 @@ class UserSearchView extends React.Component {
     data: PropTypes.object,
     onNeedMoreData: PropTypes.func,
     visibleColumns: PropTypes.arrayOf(PropTypes.string),
+    showCreateUserButton: PropTypes.bool,
     resultOffset: PropTypes.shape({
       replace: PropTypes.func.isRequired,
     }),
@@ -146,7 +149,7 @@ class UserSearchView extends React.Component {
     this.setState((state, props) => {
       const isAllChecked = !state.isAllChecked;
       const { data: { users } } = props;
-      const checkedMap = reduceUsersToMap(users, isAllChecked);
+      const checkedMap = reduceUsersToMap(users.records, isAllChecked);
 
       return {
         checkedMap,
@@ -196,6 +199,8 @@ class UserSearchView extends React.Component {
       isMultiSelect,
       resultOffset,
       initialSelectedUsers,
+      onClose,
+      showCreateUserButton,
     } = this.props;
     const { checkedMap, isAllChecked } = this.state;
 
@@ -257,6 +262,17 @@ class UserSearchView extends React.Component {
       username: user => user.username,
       email: user => get(user, ['personal', 'email']),
     };
+
+    const createUserButton = (
+      <Button
+        data-test-find-users-modal-save-multiple
+        marginBottom0
+        to={CREATE_USER_URL}
+        buttonStyle="primary"
+      >
+        <FormattedMessage id="ui-plugin-find-user.modal.button.new" />
+      </Button>
+    );
 
     return (
       <>
@@ -357,6 +373,7 @@ class UserSearchView extends React.Component {
                           paneSub={resultPaneSub}
                           defaultWidth="fill"
                           padContent={false}
+                          lastMenu={showCreateUserButton && createUserButton}
                         >
                           <MultiColumnList
                             visibleColumns={builtVisibleColumns}
@@ -405,23 +422,28 @@ class UserSearchView extends React.Component {
         {
           isMultiSelect && (
             <div className={css.UserSearchViewFooter}>
-              <>
-                <div>
-                  <FormattedMessage
-                    id="ui-plugin-find-user.modal.total"
-                    values={{ count: checkedUsersLength }}
-                  />
-                </div>
-                <Button
-                  data-test-find-users-modal-save-multiple
-                  marginBottom0
-                  onClick={this.saveMultiple}
-                  disabled={disabled}
-                  buttonStyle="primary"
-                >
-                  <FormattedMessage id="ui-plugin-find-user.modal.save" />
-                </Button>
-              </>
+              <Button
+                data-test-find-users-modal-save-multiple
+                marginBottom0
+                onClick={onClose}
+              >
+                <FormattedMessage id="ui-plugin-find-user.modal.button.cancel" />
+              </Button>
+              <div>
+                <FormattedMessage
+                  id="ui-plugin-find-user.modal.total"
+                  values={{ count: checkedUsersLength }}
+                />
+              </div>
+              <Button
+                data-test-find-users-modal-save-multiple
+                marginBottom0
+                onClick={this.saveMultiple}
+                disabled={disabled}
+                buttonStyle="primary"
+              >
+                <FormattedMessage id="ui-plugin-find-user.modal.save" />
+              </Button>
             </div>
           )
         }