+++ src/UserSearchContainer.js
@@ -40,8 +40,9 @@ const compileQuery = template(
 );
 
 export function buildQuery(queryParams, pathComponents, resourceData, logger, props) {
-  const filters = Object.keys(props.initialSelectedUsers).length ? filterConfigWithUserAssignedStatus : filterConfig;
-  const updatedResourceData = Object.keys(props.initialSelectedUsers).length && resourceData?.query?.filters?.includes(UAS) ? updateResourceData(resourceData) : resourceData;
+  const filters = props.initialSelectedUsers ? filterConfigWithUserAssignedStatus : filterConfig;
+  const updatedResourceData = props.initialSelectedUsers &&
+    resourceData?.query?.filters?.includes(UAS) ? updateResourceData(resourceData) : resourceData;
 
   return makeQueryFunction(
     'cql.allRecords=1',
@@ -159,10 +160,17 @@ class UserSearchContainer extends React.Component {
 
   onNeedMoreData = (askAmount, index) => {
     const { resultOffset } = this.props.mutator;
-
+    let offset = index;
+    if (offset < askAmount) {
+      /*
+        This condition sets offset to 100 when there are less than 100 records in the current
+        paginated result in order to skip the first 100 records and make an API call to fetch next 100.
+      */
+      offset = 100;
+    }
     if (this.source) {
-      if (resultOffset && index >= 0) {
-        this.source.fetchOffset(index);
+      if (resultOffset && offset >= 0) {
+        this.source.fetchOffset(offset);
       } else {
         this.source.fetchMore(RESULT_COUNT_INCREMENT);
       }
@@ -186,28 +194,48 @@ class UserSearchContainer extends React.Component {
       resources,
       initialSelectedUsers,
     } = this.props;
+
+    const users = {
+      records : [],
+      count: 0
+    };
     const fetchedUsers = get(resources, 'records.records', []);
     const activeFilters = get(resources, 'query.filters', '');
-    const assignedUsers = Object.values(initialSelectedUsers);
+    const assignedUsers = this.props.initialSelectedUsers ? Object.values(initialSelectedUsers) : [];
 
-    if (activeFilters === ASSIGNED_FILTER_KEY) return assignedUsers;
+    if (activeFilters === ASSIGNED_FILTER_KEY) {
+      users.records = assignedUsers;
+      users.count = assignedUsers.length;
+      return users;
+    }
 
-    if (activeFilters.includes(UAS)) {
+    if (activeFilters.includes(UAS) && this.source && this.source.loaded()) {
       const assignedUserIds = Object.keys(initialSelectedUsers);
       const hasBothUASFilters = activeFilters.includes(ASSIGNED_FILTER_KEY) && activeFilters.includes(UNASSIGNED_FILTER_KEY);
-      const hasNoneOfUASFilters = !activeFilters.includes(ASSIGNED_FILTER_KEY) && !activeFilters.includes(UNASSIGNED_FILTER_KEY);
       const uasFilterValue = activeFilters.split(',').filter(f => f.includes(UAS))[0].split('.')[1];
 
-      if (hasBothUASFilters || hasNoneOfUASFilters) {
-        return fetchedUsers;
+      if (hasBothUASFilters) {
+        users.records = fetchedUsers;
+        users.count = this.source.totalCount();
+        return users;
       }
 
       if (uasFilterValue === ASSIGNED) {
-        return fetchedUsers.filter(u => assignedUserIds.includes(u.id));
+        const filteredAssignedUsers = fetchedUsers.filter(u => assignedUserIds.includes(u.id));
+        users.records = filteredAssignedUsers;
+        users.count = filteredAssignedUsers.length;
+        return users;
       }
-      return fetchedUsers.filter(u => !assignedUserIds.includes(u.id));
+
+      const filteredUnassignedUsers = fetchedUsers.filter(u => !assignedUserIds.includes(u.id));
+      users.records = filteredUnassignedUsers;
+      users.count = this.source.totalCount() - assignedUsers.length;
+      return users;
     }
-    return fetchedUsers;
+
+    users.records = fetchedUsers;
+    users.count = this.source?.totalCount() || 0;
+    return users;
   }
 
   render() {